ルート以外のページを見られるように (#1350)
