+++ middleware.ts
@@ -1,15 +1,7 @@
 import { updateSession } from "@/lib/supabase/middleware";
-import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 
 export async function middleware(request: NextRequest) {
-  const { pathname } = request.nextUrl;
-
-  // Redirect everything except the home page and images
-  if (pathname !== "/" && !pathname.startsWith("/img/")) {
-    return NextResponse.redirect(new URL("/", request.url));
-  }
-
   return await updateSession(request);
 }
